chore(tasks): split tasks into individual invokable modules and functions
