+++ pyproject.toml
@@ -12,7 +12,7 @@ exclude = '''
     | \.tox
     | \.vscode
     | \.venv
-    | \syrupy.venv
+    | syrupy.venv
     | _build
     | buck-out
     | build
