+++ tasks/utils.py
@@ -1,0 +1,6 @@
+import os
+
+
+def ctx_run(ctx, *args, **kwargs):
+    kwargs["pty"] = os.name == "posix"
+    return ctx.run(*args, **kwargs)
